first commit of example board
This project has example code for building a robot based
on a pre-existing example board. The project has subsystems
for two motors with encoders and a single pneumatic piston.
There is a small amount of code to read values from a
limelight camera. There are three different beat commands
that use timing to generate a musical beat with the pneumatic
piston.